Move all constants to one file
the `VERSION` and `WEBSITE` constants have been moved to src/common/constants.ts
