--- src/common/version.ts
@@ -1,4 +1,0 @@
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const pjson = require('../../package.json');
-
-export const VERSION = pjson.version
