--- src/common/website.ts
@@ -1,1 +1,0 @@
-export const WEBSITE = 'https://waddleforever.com'
